Kue with redis

+++ controllers/api/v1/posts_api.js
@@ -11,8 +11,6 @@ module.exports.index = async function(req, res){
      }
    });
 
-
-
    return res.json(200,{
       message: "Lists of post",
       posts: posts
@@ -22,15 +20,22 @@ module.exports.index = async function(req, res){
 module.exports.destroy = async function(req, res){
    try{
       let post = await Post.findById(req.params.id);
-      await post.deleteOne();
-      await Comment.deleteMany({post:req.params.id});
-      return res.json(200, {
-         message: "Post and associated comments are deleted successfully!"
-      });
+      // console.log('User is ********* ', req.user.id)
+      if(post.user == req.user.id){
+         await post.deleteOne();
+         await Comment.deleteMany({post:req.params.id});
+         return res.json(200, {
+            message: "Post and associated comments are deleted successfully!"
+         });
+      }else{
+         return res.json(401 , {
+            message: 'You cannot delete this post'
+         })
+      }
    }catch(err){
       console.log("***** -ERROR- *****", err)
       return res.json(500,{
-         message: 'Internal Server Error'
+         message: 'Internal Server Error: post_api.js'
       });
    }
 }
